+++ Tests/TestProjectHelper.swift
@@ -34,8 +34,8 @@ enum TestProjectHelper {
         let project = PBXProject(
             name: name,
             buildConfigurationList: configurationList,
-            compatibilityVersion: "Xcode 14.0",
-            preferredProjectObjectVersion: 56,
+            compatibilityVersion: "Xcode 15.0",
+            preferredProjectObjectVersion: 77,
             minimizedProjectReferenceProxies: 0,
             mainGroup: mainGroup,
             developmentRegion: "en",
@@ -128,8 +128,8 @@ enum TestProjectHelper {
         let project = PBXProject(
             name: name,
             buildConfigurationList: configurationList,
-            compatibilityVersion: "Xcode 14.0",
-            preferredProjectObjectVersion: 56,
+            compatibilityVersion: "Xcode 15.0",
+            preferredProjectObjectVersion: 77,
             minimizedProjectReferenceProxies: 0,
             mainGroup: mainGroup,
             developmentRegion: "en",
@@ -221,8 +221,8 @@ enum TestProjectHelper {
         let project = PBXProject(
             name: name,
             buildConfigurationList: configurationList,
-            compatibilityVersion: "Xcode 14.0",
-            preferredProjectObjectVersion: 56,
+            compatibilityVersion: "Xcode 15.0",
+            preferredProjectObjectVersion: 77,
             minimizedProjectReferenceProxies: 0,
             mainGroup: mainGroup,
             developmentRegion: "en",